Mise au propre des states de gestionnaire Coursier
