+++ src/main/java/controleur/state/State.java
@@ -15,15 +15,15 @@ public interface State {
 
     default void modifierLivraison(StateController stateController, Livraison livraisonAModifier){}
 
-    default void ajouterCoursier(StateController stateController){}
-
-    default ArrayList<Coursier> recupererListeCoursiers(StateController stateController){
+    default ArrayList<Coursier> recupererListeCoursiers(){
         return null;
     }
 
     default void ajouterCoursier(String nom, String prenom) {}
 
     default void supprimerCoursier(Coursier coursier) {}
+
+    default int nbLivraisonAffecteCoursier(Coursier coursier) { return -1; }
 
     default void chargerLivraison(StateController stateController){}
